+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- concord-server: initial integration with Prometheus;
 - concord-server: option to disable the process state archiving task.
 
 ### Changed